@@ -1,6 +1,6 @@
 {
-  "name": "dumpsheet-webapp",
-  "version": "2.0.0",
+  "name": "merged-webapp",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "dev": "concurrently \"npm run dev:css\" \"next dev\"",
@@ -8,30 +8,57 @@
     "build": "next build",
     "build:css": "tailwindcss -m -i ./styles/tailwind.css -o styles/app.css",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "update-types": "npx supabase gen types typescript --project-id \"hzcgbsakzagvgfuektjb\" > types/supabase.ts"
   },
   "dependencies": {
+    "@hookform/resolvers": "^3.3.1",
     "@next/font": "13.1.6",
+    "@radix-ui/react-aspect-ratio": "^1.0.3",
+    "@radix-ui/react-avatar": "^1.0.3",
+    "@radix-ui/react-context-menu": "^2.1.4",
+    "@radix-ui/react-dialog": "^1.0.4",
+    "@radix-ui/react-dropdown-menu": "^2.0.5",
+    "@radix-ui/react-hover-card": "^1.0.6",
+    "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-label": "^2.0.2",
+    "@radix-ui/react-progress": "^1.0.3",
+    "@radix-ui/react-radio-group": "^1.1.3",
+    "@radix-ui/react-scroll-area": "^1.0.4",
+    "@radix-ui/react-select": "^1.2.2",
+    "@radix-ui/react-separator": "^1.0.3",
+    "@radix-ui/react-slot": "^1.0.2",
+    "@radix-ui/react-toast": "^1.1.4",
+    "@radix-ui/react-tooltip": "^1.0.6",
+    "@supabase/auth-helpers-nextjs": "^0.9.0",
     "@supabase/auth-helpers-react": "^0.3.1",
     "@supabase/auth-ui-react": "^0.2.8",
-    "@supabase/supabase-js": "^2.8.0",
-    "@types/node": "18.14.0",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.11",
-    "@vercel/analytics": "^1.1.1",
+    "@supabase/supabase-js": "^2.40.0",
+    "@types/node": "20.3.1",
+    "@types/react": "18.2.12",
+    "@types/react-dom": "18.2.5",
+    "@vercel/analytics": "^1.2.2",
+    "@vercel/speed-insights": "^1.0.12",
+    "autoprefixer": "^10.4.20",
+    "axios": "^1.6.8",
+    "class-variance-authority": "^0.7.0",
+    "disposable-email-domains": "^1.0.62",
     "eslint": "8.34.0",
     "eslint-config-next": "13.1.6",
-    "next": "^13.1.6",
+    "lucide-react": "^0.436.0",
+    "next": "^14.1.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "recordrtc": "^5.6.2",
-    "typescript": "4.9.5"
+    "stripe": "^16.9.0",
+    "tailwind-merge": "^2.5.2",
+    "tailwindcss": "^3.3.2",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "5.1.3"
   },
   "devDependencies": {
     "@types/recordrtc": "^5.6.11",
     "concurrently": "^7.6.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "tailwindcss": "^3.2.7",
-    "vercel": "^32.3.1"
+    "eslint-plugin-react-hooks": "^4.6.0"
   }
 }
