@@ -1,47 +1,15 @@
-import {useRouter} from "next/router";
+"use client"
+import {useSearchParams} from "next/navigation";
 import Image from "next/image";
 
-type BannerProps = {
-    showExtras: boolean;
-};
+type BannerProps = {};
 
-const Banner: React.FC<BannerProps> = ({ showExtras }) => {
-    const router = useRouter();
-    const {hubspot_status, account_id} = router.query;
+const Banner: React.FC<BannerProps> = ({  }) => {
+    const searchParams = useSearchParams()
+    const hubspot_status = searchParams.get('hubspot_status');
 
     return (
         <>
-            <div className="absolute top-6 left-4 flex items-center space-x-2 top-navigation">
-                <a href="https://www.dumpsheet.com/">
-                    <Image
-                        src="/images/dumpsheet-logo-transparent.png"
-                        alt="Dumpsheet Logo"
-                        width={150}
-                        height={30}
-                    />
-                </a>
-                {/*
-                  <div className="flex items-center space-x-8 font-medium text-black text-lg tracking-tight">
-                      <div>Features</div>
-                      <div>Use cases</div>
-                  </div>
-                  */}
-            </div>
-            <div className="absolute top-3 right-4 flex items-center space-x-2">
-                <a href="https://www.loom.com/share/1614e907aeea4312bb53affd99677593" target="_blank"
-                   rel="noopener noreferrer">
-                    <button
-                        className="flex items-center justify-center w-40 h-12 bg-black rounded-full font-semibold text-white text-lg tracking-tighter hover:bg-gray-700">
-                        Watch Demo
-                    </button>
-                </a>
-                {/*
-                <div className="flex items-center justify-center w-12 h-12 bg-[#0000001a] rounded-full">
-                    <Image className="group-8" alt="Group" src="/images/figma/group-1000004834.png" width={30}
-                           height={30}/>
-                </div>
-                */}
-            </div>
             {hubspot_status && (
                 <div className="absolute top-20 left-1/2 transform -translate-x-1/2 flex flex-col items-center justify-center text-center border border-black lg bg-white px-6 py-1">
                     <div className="block text-black font-semibold text-center">
@@ -53,7 +21,7 @@ const Banner: React.FC<BannerProps> = ({ showExtras }) => {
                     </div>
                 </div>
             )}
-            {!hubspot_status && showExtras && (
+            {!hubspot_status && (
                 <div className="w-[16rem] absolute top-20 left-1/2 transform -translate-x-1/2 flex flex-col items-center justify-center text-center border border-black rounded-lg bg-white px-6 py-1">
                     <div className="block text-black font-semibold text-center">
                         We are in Private Beta
