@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-}
+  experimental: {
+    esmExternals: "loose",
+  },
+};
 
 module.exports = nextConfig
